simplifying node selector/toleration submission example
